OpenP5/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.21.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.1.1=pypi_0
annotated-types=0.5.0=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
bitsandbytes=0.39.0=pypi_0
black=23.7.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.11.17=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
contourpy=1.1.0=pypi_0
cryptography=41.0.2=py39h774aba0_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pypi_0
datasets=2.14.4=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.7=pypi_0
evaluate=0.4.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pypi_0
fastapi=0.103.1=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpy=0.3.1=pypi_0
filelock=3.12.3=pypi_0
fire=0.5.0=pypi_0
fonttools=4.42.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.9.0=pypi_0
fuzzywuzzy=0.18.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gradio=3.42.0=pypi_0
gradio-client=0.5.0=pypi_0
grpcio=1.57.0=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h11=0.14.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.22.2=pypi_0
idna=3.4=py39h06a4308_0
imageio=2.31.4=py39h06a4308_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.15.0=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
levenshtein=0.25.1=pypi_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
loralib=0.1.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.24.3=py39h14f4228_0
numpy-base=1.24.3=py39h31eccc5_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
orjson=3.9.5=pypi_0
packaging=23.1=pypi_0
pandas=2.1.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.2=pypi_0
peft=0.2.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.2.1=py39h06a4308_0
platformdirs=3.10.0=pypi_0
pluggy=1.3.0=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.24.2=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py3nvml=0.2.7=pypi_0
pyarrow=13.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.3.0=pypi_0
pydantic-core=2.6.3=pypi_0
pydub=0.25.1=pypi_0
pygments=2.16.1=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39h06a4308_0
pytest=7.4.1=pypi_0
python=3.9.0=hdb3f193_2
python-dateutil=2.8.2=pypi_0
python-levenshtein=0.25.1=pypi_0
python-multipart=0.0.6=pypi_0
pytorch=1.13.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
rapidfuzz=3.8.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
regex=2023.8.8=pypi_0
requests=2.31.0=py39h06a4308_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rpds-py=0.10.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.2=pypi_0
semantic-version=2.10.0=pypi_0
sentence-transformers=2.7.0=pypi_0
sentencepiece=0.1.99=pypi_0
setproctitle=1.2.2=py39h27cfd23_1004
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
starlette=0.27.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorboardx=2.2=pyhd3eb1b0_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenize-rt=5.2.0=pypi_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
torchaudio=0.13.1=py39_cu117
torchvision=0.14.1=py39_cu117
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.28.0=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=py39h06a4308_0
uvicorn=0.23.2=pypi_0
wcwidth=0.2.6=pypi_0
websockets=11.0.3=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py39h06a4308_0
xmltodict=0.13.0=pypi_0
xxhash=3.3.0=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
